fix: patch remaining validators for cryptographic sorting
- Injected `sort_arrays` validator to `WorkingMemorySnapshot` to ensure determinism across `theory_of_mind_models` and `capability_attestations`.
- Added `sort_adapters` validator to `AgentNode`.
- Added `sort_arrays` validator to `FederatedDiscoveryProtocol`.
- Overwrote `WorkingMemorySnapshot` fields to strip lazy evaluation quotes that could crash `__setattr__`.
- Sorted arrays in `test_working_memory_theory_of_mind` to reflect the newly mandated canonical form and satisfy strict `frozen=True` mutations.
